feat(lab): three-arm Harvey LAB (Opus A/B + Nemotron C)
Wire Arm C via clawql-cc OpenRouter chat adapter, share vault/MCP
session code, extend GHA arms, and refresh ledgers/handoff for the
August 2026 Nemotron post-train strategy (skip blocking fine-tune).
