+++ docs/benchmarks/README.md
@@ -18,10 +18,11 @@ This directory holds **repro instructions**, **latest pointers**, **per-scenario
 | [`openbench-advanced-suites.md`](openbench-advanced-suites.md)                                             | **Next-gen suites (B-1…B-7)** broken into small tasks, phases, and gates.                                                                                        |
 | [`openbench-advanced-specs.md`](openbench-advanced-specs.md)                                               | Advanced suites B-1…B-7 (specs; Phase 1 offline packs).                                                                                                          |
 | [`openbench-b7-calderwood.md`](openbench-b7-calderwood.md)                                                 | **B-7** C&H institutional knowledge + amortization. Essay: [Memory Finds. Ontology Decides.](https://pragmaticvectors.com/posts/memory-finds-ontology-decides/). |
-| [`harvey-lab-baseline.md`](harvey-lab-baseline.md)                                                         | Harvey LAB **firm-knowledge** baseline (standard harness).                                                                                                       |
-| [`harvey-lab-clawql-results.md`](harvey-lab-clawql-results.md)                                             | Harvey LAB × ClawQL two-arm ledger (Opus vs Opus when complete).                                                                                                 |
-| [`harvey-lab-pause-handoff.md`](harvey-lab-pause-handoff.md)                                               | **Pause/resume** for LAB × ClawQL (GHA + OpenRouter next steps).                                                                                                 |
-| [`../inference/clawql-inference-training-pipeline.md`](../inference/clawql-inference-training-pipeline.md) | Post-LAB fine-tune flywheel (SFT → DPO → GRPO → SPIN).                                                                                                           |
+| [`harvey-lab-baseline.md`](harvey-lab-baseline.md)                                                         | Harvey LAB **firm-knowledge** baseline (Arm A, standard harness).                                                                                                |
+| [`harvey-lab-clawql-results.md`](harvey-lab-clawql-results.md)                                             | Harvey LAB × ClawQL three-arm ledger (Opus A/B + Nemotron C when complete).                                                                                      |
+| [`harvey-lab-pause-handoff.md`](harvey-lab-pause-handoff.md)                                               | **Pause/resume** for LAB × ClawQL (GHA + OpenRouter + Arm C).                                                                                                    |
+| [`harvey-lab-action-plan.md`](harvey-lab-action-plan.md)                                                   | Reconciles the August 2026 Cursor action plan with shipped overlay.                                                                                              |
+| [`../inference/clawql-inference-training-pipeline.md`](../inference/clawql-inference-training-pipeline.md) | Post-LAB fine-tune flywheel (optional after ledger; Arm C skips blocking fine-tune).                                                                             |
 | [`ouroboros-value-evidence.md`](ouroboros-value-evidence.md)                                               | **Evidence:** Ouroboros on converges vs off strategy thrash (verified A/B).                                                                                      |
 | [`openbench-ouroboros-oscillation.md`](openbench-ouroboros-oscillation.md)                                 | Ouroboros oscillation-escape task design, caps, repro.                                                                                                           |
 | [`openbench-github-actions.md`](openbench-github-actions.md)                                               | One-off Actions A/B (clawql-on vs off) — spin up, report, spin down.                                                                                             |
