+++ docs/benchmarks/harvey-lab-baseline.md
@@ -1,27 +1,28 @@
 # Harvey LAB Baseline — firm-knowledge
 
-Date: 2026-08-08  
-Model: **pending** (target publishable: `claude-opus-4-6`; debug: `claude-sonnet-4-6`)  
+Date: 2026-08-11  
+Model: **pending** (publishable Arm A: `claude-opus-4-8`; debug: `claude-sonnet-4-6`)  
 Tasks: **250** (`tasks/firm-knowledge/tasks/*/task.json`)  
 Shared DMS: ~266 matters / ~9,288 files / ~525 MB (`tasks/firm-knowledge/dms`)  
 Judge: `claude-sonnet-4-6`  
-Notes: standard harness, no ClawQL
+Notes: standard harness, no ClawQL (Arm A)
 
 ## Status
 
-**Paused — resume in GitHub Actions** with repo secret `OPENROUTER_API_KEY` (same as OpenBench).  
-See [`harvey-lab-pause-handoff.md`](harvey-lab-pause-handoff.md).
+**Ready to score in GitHub Actions** with repo secret `OPENROUTER_API_KEY`.  
+Three-arm strategy: [`harvey-lab-pause-handoff.md`](harvey-lab-pause-handoff.md).  
+Action-plan reconciliation: [`harvey-lab-action-plan.md`](harvey-lab-action-plan.md).
 
 Harness readiness completed without scoring:
 
-| Check                                                | Result                                                                   |
-| ---------------------------------------------------- | ------------------------------------------------------------------------ |
-| Clone `harveyai/harvey-labs`                         | OK                                                                       |
-| Docs (`tutorial`, `architecture`, `eval-strategies`) | Read                                                                     |
-| `firm-knowledge` inventory                           | 250 tasks; shared DMS via `docs_dir: ../../dms`                          |
-| First task                                           | `firm-knowledge/tasks/001` — Antitrust HSR Second Requests (11 criteria) |
-| `uv sync` / Podman / sandbox                         | OK (agent smoke)                                                         |
-| Inference path                                       | **GHA** workflow `harvey-lab-firm-knowledge.yml` + OpenRouter            |
+| Check | Result |
+| ----- | ------ |
+| Clone `harveyai/harvey-labs` | OK |
+| Docs (`tutorial`, `architecture`, `eval-strategies`) | Read |
+| `firm-knowledge` inventory | 250 tasks; shared DMS via `docs_dir: ../../dms` |
+| First task | `firm-knowledge/tasks/001` — Antitrust HSR Second Requests |
+| `uv sync` / Podman / sandbox | OK (agent smoke) |
+| Inference path | **GHA** `harvey-lab-firm-knowledge.yml` + OpenRouter |
 
 ## Criterion pass rate
 
@@ -35,7 +36,7 @@ _Not yet measured._
 
 _Empty until Phase A Sonnet single-task and Phase E Opus sweep complete._
 
-## Next (on resume)
+## Next
 
-1. `gh workflow run harvey-lab-firm-knowledge.yml` Phase A (`tasks/001`, Sonnet, both arms).
-2. Record CPR / all-pass here for Opus baseline (same model as ClawQL arm).
+1. `gh workflow run harvey-lab-firm-knowledge.yml` Phase A (`tasks/001`, Sonnet, `baseline,clawql`).
+2. Record CPR / all-pass here for Opus Arm A (same model as Arm B).
