+++ integrations/harvey-labs/harness/adapters/clawql_openrouter.py
@@ -3,13 +3,14 @@
 When ``CLAWQL_LAB_USE_OPENROUTER=1`` (or Anthropic key is absent and
 ``OPENROUTER_API_KEY`` is set), Anthropic SDK calls go through
 ``https://openrouter.ai/api`` with the OpenRouter key.
+
+Chat-completions (Nemotron Arm C) always use the OpenAI SDK against OpenRouter.
 """
 
 from __future__ import annotations
 
 import os
-
-import anthropic
+from typing import Any
 
 # Harvey short ids → OpenRouter Anthropic model ids.
 _OPENROUTER_MODEL_MAP = {
@@ -23,6 +24,14 @@ _OPENROUTER_MODEL_MAP = {
     "claude-opus-4.8": "anthropic/claude-opus-4.8",
 }
 
+# Short aliases → OpenRouter chat-completions model ids (Arm C).
+_OPENROUTER_CHAT_MODEL_MAP = {
+    "nemotron-3.5-lightning": "nvidia/nemotron-3.5-lightning:free",
+    "nemotron-3.5-lightning:free": "nvidia/nemotron-3.5-lightning:free",
+    "nvidia/nemotron-3.5-lightning": "nvidia/nemotron-3.5-lightning:free",
+    "nvidia/nemotron-3.5-lightning:free": "nvidia/nemotron-3.5-lightning:free",
+}
+
 
 def use_openrouter() -> bool:
     if os.environ.get("CLAWQL_LAB_USE_OPENROUTER", "").strip().lower() in {
@@ -31,14 +40,13 @@ def use_openrouter() -> bool:
         "yes",
     }:
         return bool(os.environ.get("OPENROUTER_API_KEY"))
-    # Fall back to OpenRouter when no direct Anthropic key (GHA OpenBench pattern).
     return bool(os.environ.get("OPENROUTER_API_KEY")) and not os.environ.get(
         "ANTHROPIC_API_KEY"
     )
 
 
 def resolve_openrouter_model(model: str) -> str:
-    """Map a Harvey / short Claude id to an OpenRouter model id."""
+    """Map a Harvey / short Claude id to an OpenRouter Anthropic model id."""
     if model.startswith("anthropic/"):
         return model
     if model.startswith("openrouter/"):
@@ -46,29 +54,70 @@ def resolve_openrouter_model(model: str) -> str:
     return _OPENROUTER_MODEL_MAP.get(model, f"anthropic/{model}")
 
 
-def make_anthropic_client() -> anthropic.Anthropic:
+def resolve_openrouter_chat_model(model: str) -> str:
+    """Map Arm C short ids to OpenRouter chat-completions model ids."""
+    if model.startswith("openrouter/"):
+        return model.removeprefix("openrouter/")
+    mapped = _OPENROUTER_CHAT_MODEL_MAP.get(model)
+    if mapped:
+        return mapped
+    if model in {"nemotron", "nemotron-lightning"}:
+        return os.environ.get(
+            "CLAWQL_LAB_NEMOTRON_MODEL",
+            "nvidia/nemotron-3.5-lightning:free",
+        )
+    return model
+
+
+def _openrouter_headers() -> dict[str, str]:
+    return {
+        "HTTP-Referer": os.environ.get(
+            "CLAWQL_OPENROUTER_HTTP_REFERER", "https://clawql.com"
+        ),
+        "X-Title": os.environ.get(
+            "CLAWQL_OPENROUTER_APP_TITLE", "ClawQL Harvey LAB"
+        ),
+    }
+
+
+def make_anthropic_client() -> Any:
     """Anthropic SDK client — direct or via OpenRouter."""
+    import anthropic
+
     if use_openrouter():
         return anthropic.Anthropic(
             api_key=os.environ["OPENROUTER_API_KEY"],
             base_url=os.environ.get(
                 "CLAWQL_LAB_OPENROUTER_BASE_URL", "https://openrouter.ai/api"
             ),
-            default_headers={
-                "HTTP-Referer": os.environ.get(
-                    "CLAWQL_OPENROUTER_HTTP_REFERER", "https://clawql.com"
-                ),
-                "X-Title": os.environ.get(
-                    "CLAWQL_OPENROUTER_APP_TITLE", "ClawQL Harvey LAB"
-                ),
-            },
+            default_headers=_openrouter_headers(),
             max_retries=1,
         )
     return anthropic.Anthropic(max_retries=1)
 
 
+def make_openrouter_openai_client() -> Any:
+    """OpenAI SDK client pointed at OpenRouter (Chat Completions / Arm C)."""
+    import openai
+
+    key = os.environ.get("OPENROUTER_API_KEY")
+    if not key:
+        raise ValueError(
+            "OPENROUTER_API_KEY required for clawql-cc / Nemotron Arm C"
+        )
+    return openai.OpenAI(
+        api_key=key,
+        base_url=os.environ.get(
+            "CLAWQL_LAB_OPENROUTER_OPENAI_BASE_URL",
+            "https://openrouter.ai/api/v1",
+        ),
+        default_headers=_openrouter_headers(),
+        max_retries=1,
+    )
+
+
 def maybe_rewrite_model(model: str) -> str:
-    """Rewrite model id for OpenRouter when that path is active."""
+    """Rewrite Anthropic model id for OpenRouter when that path is active."""
     if use_openrouter():
         return resolve_openrouter_model(model)
     return model
