+++ integrations/harvey-labs/harness/clawql_tools.py
@@ -1,15 +1,21 @@
-"""ToolExecutor wrapper that routes clawql_* tools to the ClawQL adapter."""
+"""ToolExecutor wrapper that routes clawql_* tools to the ClawQL LAB session."""
 
 from __future__ import annotations
 
-from harness.adapters.clawql import ClawQLAdapter, _TOOL_NAME_TO_MCP
+from typing import Any, Protocol
+
+from harness.adapters.clawql_lab_session import _TOOL_NAME_TO_MCP
 from harness.tools import ToolExecutor
+
+
+class _ClawQLToolHost(Protocol):
+    def execute_clawql_tool(self, tool_name: str, arguments: str | dict) -> str: ...
 
 
 class ClawQLToolExecutor(ToolExecutor):
     """Sandbox tools + ClawQL MCP tools."""
 
-    def __init__(self, clawql_adapter: ClawQLAdapter, **kwargs):
+    def __init__(self, clawql_adapter: _ClawQLToolHost, **kwargs: Any):
         super().__init__(**kwargs)
         self.clawql_adapter = clawql_adapter
         self.clawql_calls: int = 0
